requirements.txt
torch
torchvision
numpy
tqdm
tensorboard
tensorboardX
ml-collections
medpy
SimpleITK
scipy
h5py
timm
einops
imagecodecs
matplotlib
seaborn
opencv-python
pillow
tifffile
